some minor fixes

+++ SRC/MCAL/STK/stk.c
@@ -75,7 +75,7 @@ void SysTick_Handler(void){
 uint8_t stk_setTime_mS(uint32_t time_mS) {
     uint8_t errorStatus = STK_NOT_OK;
     if(time_mS <= STK_MAX_TIME_MS){
-        uint32_t counts = ((uint32_t)time_mS*CLK/1000) - 1;
+        uint32_t counts = ((uint32_t)(uint64_t)time_mS*CLK/1000) - 1;
         stk->LOAD &= ~LOAD_VAL_MASK;
         stk->LOAD |= (counts & LOAD_VAL_MASK);
         stk->VAL &= ~CURRENT_VAL_MASK;
